Worked on ShooterTestCommand #43
I am not done yet still. I just want you to look at this because everything was appearing on shuffleboard, and all of the sudden, now it's not.  If you can figure out why that would help.

I changed MotorPlacement to public, I don't know if I was supposed to do that, but I used it in the TestCommand.